services/mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "1.0.0"
description = "MCP Server - Central state management API for Agentic FEA workflows"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "azure-storage-blob>=12.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []

